+++ src/elements/index.ts
@@ -1,3 +1,3 @@
 // Export all elements from here
 export { default as Button } from './button';
-export { default as Checkbox } from './checkboxrenaming';
+export { default as Checkbox } from './Checkbox';
